Uploaded eSign+ itself (original, found)
This is the first version of the unmodified eSign+ program. This program has been found from an archive of a now-defunct URL that has been serving the tarred eSign+ program. It hasn't been modified and this is simply the app itself, used within eSign+'s helper app (more information can be found on: https://techybuff.com/resources) to install the application itself. This tar has remained unfound for a long time in Reddit threads and only YouTube video with the now-defunct URL has covered it (with recent comments saying it doesn't work). More archives are to be published soon, the eSign+ ZIP to be published soon (found on https://techybuff.com/resources) and also the improved eSign+ version will be uploaded soon.