Massive README.md update
This is the README.md with some common questions and things for users. A good chunk of it is not included including a TODO section in the file referred to by another actually existing section. The current TODOs for this file is to improve it and better it's markup (formatting).