Massive update to README.md
This massive update to README.md rewrites an entire section for the better and also starts writing about setting up eSign+ and installing it. It also includes all the methods you need to know, however the "Traditional Method" is the first one being written and it's not finished yet. In other words, the README.md is slightly better and the section of setting the app up is being written.

Signed-off-by: EvilDagger <evildagger@proton.me>