Another small logic fix on README.md
Before the description, using Preview wasn't utilized for the README.md (or anything in the past), now it is mandatory and recommended to do so for this organisation and repository. This fixes a small logic fix which is very important to markup and presentation, for further reminder: add newline after using x.x numbering scheme when using x numbering scheme after it!

Signed-off-by: EvilDagger <evildagger@proton.me>