Uploads the actual working and tested IPA file
The past IPA files were corrupted and unfit however this IPA file has been tested for TrollStore usage and general usage and it works. This is an exciting new leap for eSign+ since it brings it to alpha phase the very second this commit is done, makes the application better and apart from that, 6.73333333333x easier and friendlier to setup using a new method with 3 steps, not 20.2 like the first method!

Signed-off-by: EvilDagger <evildagger@proton.me>